services/api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
prometheus-client==0.19.0
nats-py==2.6.0
redis==5.0.1